+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fixed an issue where Twig macros couldn't be imported into a No-Cache block from the template the No-Cache block is a part of
+
 ## 2.0.7 - 2020-12-31
 
 ### Fixed